(feat) latoken connector integration + 100% unit test coverage
