--- hummingbot/connector/exchange/latoken/latoken_ws_connection.py
@@ -1,8 +1,0 @@
-from hummingbot.core.web_assistant.connections.data_types import WSRequest
-from hummingbot.core.web_assistant.connections.ws_connection import WSConnection
-
-
-class LatokenWSConnection(WSConnection):
-    async def send(self, request: WSRequest):
-        self._ensure_connected()
-        await self._connection.send_str(request.payload)
